refactor: 인터페이스 및 구현체 클래스 이름 변경
