--- src/main/java/roomescape/dao/ReservationDaoInterface.java
@@ -1,15 +1,0 @@
-package roomescape.dao;
-
-import java.util.List;
-import roomescape.entity.ReservationWithTimeId;
-import roomescape.model.Reservation;
-
-public interface ReservationDaoInterface {
-
-    List<Reservation> selectAllReservation();
-
-    Long addReservation(ReservationWithTimeId reservation);
-
-    int deleteReservationById(Long id);
-
-}
